start refactoring VM into separate type with methods.
I couldn't get this to run fast in my previous experiments but apparently this is due to the IP and BP not being on the VM type itself but stored in the current frame (last element in the Vec<Frame> property). If we copy these over to the VM type things are fast (if not faster!) than one giant function.