Run the GC on function return
Even with GC enabled on function returns, which is exercised
extensively with the fib(24) benchmark, this makes the runtime only
around 1% slower.  Compare with the unmodified GC, also being
executed on function returns, that made fib(25) 25% slower.

The parser benchmark is around 20% slower, though, so I got to look
into it.